% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg-apply.R
\name{sg_apply}
\alias{sg_apply}
\title{Combinatory Savitzky-Golay Filtering}
\usage{
sg_apply(
  X,
  dt_sg_plan,
  dt_prep_sets = NULL,
  nest_params = TRUE,
  append_rows = FALSE
)
}
\arguments{
\item{X}{\code{matrix}, \code{data.frame} or \code{data.table} with spectra to be
preprocessed according to plan (see \code{dt_sg_plan}).}

\item{dt_sg_plan}{A standardized \code{data.table} with the Savitzky-Golay
parameter sets, which can be generated with \code{sg_plan()}. It must at least
contain the following columns:
\itemize{
\item \code{prep_set} (character)
\item \code{prep_label} (character)
\item \code{m} (integer): order of the derivative; \code{m = 0} signifies no derivative
\item \code{p} (integer): polynomal order
\item \code{w} (integer): window size in number of spectral points; must be uneven
\code{m} is the , \code{p} is the polynomial order
that should be bigger than the derivative order,  and \code{w} is the window size
in  number of spectral points (must be uneven). See section
\emph{Savitzky-Golay Plan} for templating the required object and
\code{\link[prospectr:savitzkyGolay]{prospectr::savitzkyGolay()}} for further the original Savitzky-Golay
algorithm.
}}

\item{dt_prep_sets}{A standardized \code{data.table}, i.e. returned from \verb{specprep::*_apply()}
function. Contains labelled sets of preprocessed spectra:
This argument allows to chain preprocessing in sequential manner, and i.e. apply variable
Savitzky-Golay smoothers with a single function application.}

\item{nest_params}{tbd}

\item{append_rows}{tbd}
}
\value{
data.table with the following (list)columns:by
#to be filled
}
\description{
Apply Savitzky-Golay filtering at variable combinations of
parameter sets for set(s) of spectra.
}
\details{
\subsection{Design principles}{

Savitzky-Golay transformation (moving window polynomial least-squares) prior modeling can help
to reduce noise and enhance signals in spectra. This can allowing models to extract parsimonious
predictable information from spectra for more accurate estimation. However, this process requires
empirical optimization and fine-tuning of the parameters that control the nature and degree of
smoothing and hence noise removal for calibration task at hand, which is often not done. For
example, systematically varying the size of the smoothing window control the amount of
information filtered and potential artefacts created. Nonetheless, non-stationary noise as
opposed to white gaussian noise and informative fluctuations in chemically-driven spectral
dynamics (e.g. slope changes and different absorption peak widths and compositional complexity)
can make a simple nonrecursive application of the original Savitzky-Golay algorithm less
appropriate to filter noise.

Templating code for sequential and/or recursive branching of preprocessing methods with variation
their parameters, if applicable, can be repetitive and cumbersome. This is where the specprep
package with combinatory planning and application tools jumps in.

The combinatory power of the \code{sg_apply()} function stems from the ability to map Savitzly Golay
both over row-wise sets of parametrizations (see subsection \emph{Savitzky-Golay Plan}) and previous
preprocessing rounds that yielded set(s) of (differently) processed spectra to be processed again
(see section \emph{Set(s) of Previously Processed Spectra}). Since \code{data.table}s are structured
consistently across the \verb{specprep::*_apply} type of functions, their inputs and outputs are
interoperable. This allows flexiblity for applying combinations of preprocessing methods.
}

\subsection{Savitzky-Golay Plan}{

\code{dt_sg_plan} is most conventiently built with \code{sg_plan()}. It parametrizes Savitzky-Golay
preprocessing scheduled on either \code{X} or on all sets of already processed spectra contained in
\code{dt_prep_sets}. Each row lays out one preprocessing step, linking the following data across
columns:
\itemize{
\item \code{prep_set}: this string identifies the name of general preprocessing method that is chained
to sets of spectra.
}
}

\subsection{Set(s) of Previously Processed Spectra}{

tbd
}
}
\author{
Philipp Baumann
}
