% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg-apply.R
\name{sg_make_plan}
\alias{sg_make_plan}
\title{Generate a data.frame with Savitzky-Golay parameters}
\usage{
sg_make_plan(param_list)
}
\arguments{
\item{param_list}{A list of}
}
\value{
data.frame
}
\description{
Make a full-factorial combination of Savitzky-Golay parameters.
}
\author{
Philipp Baumann
}
