% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv-apply.R
\name{snv_apply}
\alias{snv_apply}
\title{Compute the standard normal variate for collections of spectra}
\usage{
snv_apply(X, dt_prep_sets = NULL, append_rows = FALSE)
}
\arguments{
\item{X}{\code{matrix}, \code{data.frame} or \code{data.table} with spectra used as input
to compute standard normal variate (SNV)}

\item{dt_prep_sets}{A standardized \code{data.table}, i.e. returned from \verb{specprepper::*_apply()}
function. Contains labelled sets of preprocessed spectra:
This argument allows to chain preprocessing in sequential manner, and i.e. apply variable
Savitzky-Golay smoothers with a single function application.}

\item{append_rows}{logical whether to append the newly processed rows, when
\code{dt_prep_sets} is not NULL.}
}
\description{
Compute the standard normal variate for collections of spectra
}
