% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colmean-group-apply.R
\name{colmean_group_apply}
\alias{colmean_group_apply}
\title{Compute the mean spectra per group label for all spectra in a collection}
\usage{
colmean_group_apply(dt_prep_sets, append_rows = FALSE)
}
\arguments{
\item{dt_prep_sets}{A standardized \code{data.table}, i.e. returned from \verb{specprepper::*_apply()}
function. Contains labelled sets of preprocessed spectra:
This argument allows to chain preprocessing in sequential manner, and i.e. apply variable
Savitzky-Golay smoothers with a single function application.}

\item{append_rows}{logical whether to append the newly processed rows, when
\code{dt_prep_sets} is not NULL.}
}
\description{
Compute the mean spectra per group label for all spectra in a collection
}
